requirements.txt
discord.py
python-dotenv
openai
coverage
codecov

kluvs-brain @ git+https://github.com/ivangarzab/kluvs-brain.git
